First/Last HTML and CSS Update.
Changed the HTMl file to index.html. Changed the CSS file from ss_layout_txt.css to ss_layout.css. The rule is for the first line of the first paragraph and the has small caps on the first line of the paragraph. This rule is for the first letter of the first paragraph and its floated left, line height of .8em, the margins are right/bottom 5 pixels. The rule is for the interactive map and its in the div that has a id of battleMap. the positioning is relative and the top bottom margins are twenty pixels. by doing so, causes the map to center. This rule is to make the points on the map to hover and to have the information to show up when the mouse hovers the certain points. ready for review.